Refactor cmdVersion to cli package
